+++ CHANGELOG.md
@@ -30,6 +30,20 @@
 - [v0.1.1](#v011)
 - [v0.1.0](#v010)
 
+## Unreleased
+
+### Added
+
+- Support specifying labels to filter watched `Secret`s and `ConfigMap`s of
+  each `ControlPlane` by `spec.objectFilters.secrets.matchLabels` and
+  `spec.objectFilters.configMaps.matchLabels`. Only secrets or configmaps that
+  have the labels matching the specified labels in spec are reconciled.
+  If Kong operator has also flags `--secret-label-selector` or
+  `--config-map-label-selector` set, the controller for each `ControlPlane` also
+  requires reconciled secrets or configmaps to set the labels given in the flags
+  to `true`.
+  [#1982](https://github.com/Kong/kong-operator/pull/1982)
+
 ## [v2.0.0-alpha.3]
 
 > Release date: 2025-08-08
@@ -146,6 +160,9 @@
   by introducing the `ExtensionProcessor` interface.
   This change enables KonnecExtensions for `ControlPlane v2alpha1`.
   [#1978](https://github.com/Kong/kong-operator/pull/1978)
+
+### Changes
+
 - `ControlPlane` provisioned conditions' reasons have been renamed to actually reflect
   the new operator architecture. `PodsReady` is now `Provisioned` and `PodsNotReady`
   is now `ProvisioningInProgress`.